+++ src/gui.js
@@ -10063,11 +10063,9 @@ CostumeIconMorph.prototype.removeCostume = function () {
 };
 
 CostumeIconMorph.prototype.importEmbeddedData = function () {
-    this.parentThatIsA(IDE_Morph).droppedText(
-        this.object.embeddedData,
-        this.object.name,
-        ''
-    );
+    var ide = this.parentThatIsA(IDE_Morph);
+    ide.spriteBar.tabBar.tabTo('scripts');
+    ide.droppedText(this.object.embeddedData, this.object.name, '');
 };
 
 CostumeIconMorph.prototype.typeOfStringData = function (aString) {